+++ CHANGES.rst
@@ -6,6 +6,9 @@ Changelog
 
 - Override folderish viewlet (method available) to avoid printing data structure in folderish element view.
   [boulch]
+ 
+- uncomment IUseKeywordHomepage interface (Florenne "hack" for linked contact).
+  [boulch]
 
 
 0.13.16 (2019-07-22)